* Framework/PCProjectBrowser.m         * Framework/PCSaveModified.m         * Framework/PCProjectLoadedFiles.m         * Framework/PCProjectBuilder.m         * Headers/ProjectCenter/PCProjectBrowser.h         * Headers/ProjectCenter/PCSaveModified.h         * Headers/ProjectCenter/PCProjectLoadedFiles.h         * Modules/Projects/Application/PCAppProject+Inspector.h         * Modules/Projects/Application/PCAppProject+Inspector.m           some more (unsigned) int -> NS(U)Integer transitions           not catched from clang, but found by libobjc2 in debug mode
OK Fred



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/projectcenter/trunk@36108 72102866-910b-0410-8b05-ffd578937521

+++ Framework/PCProjectBrowser.m
@@ -527,13 +527,13 @@ NSString *PCBrowserDidSetPathNotification = @"PCBrowserDidSetPathNotification";
 @implementation PCProjectBrowser (ProjectBrowserDelegate)
 
 - (void)     browser:(NSBrowser *)sender
- createRowsForColumn:(int)column
+ createRowsForColumn:(NSInteger)column
 	    inMatrix:(NSMatrix *)matrix
 {
-  NSString  *pathToCol;
-  NSArray   *files;
-  int       i = 0;
-  int       count = 0;
+  NSString   *pathToCol;
+  NSArray    *files;
+  NSUInteger i = 0;
+  NSUInteger count = 0;
 
   if (sender != browser || !matrix || ![matrix isKindOfClass:[NSMatrix class]])
     {
