+++ Headers/ProjectCenter/PCSaveModified.h
@@ -55,11 +55,11 @@ BOOL PCRunSaveModifiedFilesPanel(PCEditorManager *manager,
 - (BOOL)saveSelectedFiles;
 
 // TableView delegate
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView;
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView;
 
 - (id)            tableView:(NSTableView *)aTableView
   objectValueForTableColumn:(NSTableColumn *)aTableColumn
-                        row:(int)rowIndex;
+                        row:(NSInteger)rowIndex;
 @end
